using long for user id

+++ Sadie.Networking.Writers/Rooms/Users/RoomUserEffectWriter.cs
@@ -7,13 +7,13 @@ namespace Sadie.Networking.Writers.Rooms.Users;
 [PacketId(ServerPacketId.RoomUserEffect)]
 public class RoomUserEffectWriter : AbstractPacketWriter
 {
-    public required int UserId { get; init; }
+    public required long UserId { get; init; }
     public required int EffectId { get; init; }
     public required int DelayMs { get; init; }
 
     public override void OnSerialize(INetworkPacketWriter writer)
     {
-        writer.WriteInteger(UserId);
+        writer.WriteLong(UserId);
         writer.WriteInteger(EffectId);
         writer.WriteInteger(DelayMs);
     }
